Move FadedScrollView
