+++ generatedTypes/src/components/fadedScrollView/index.d.ts
