+++ src/components/fadedScrollView/index.tsx
